HS-1074 fix unit tests
